backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6
slowapi==0.1.9
python-dotenv==1.0.0
